+++ prompter_app/src/routes/PromptBox.svelte
@@ -1,18 +1,15 @@
 <script lang="ts">
   import { onMount } from 'svelte';
-  import type { Prompt } from '$lib/prompts';
+  import { promptSchemaVersion, type Prompt } from '$lib/prompts';
   import { renderPrompt } from '$lib/prompts';
   import { escapeHtml } from '$lib/util';
-  import { page } from '$app/stores';
 
   import '$lib/codemirror5/codemirror.css';
 
   // Model
-	export let promptText = "Tell me a story about {{ storyTopic }}, make it sound like you're very excited about {{ storyTopic | title }}!\n\n{% if anotherTopic %}\nAnd another one about {{ anotherTopic | upper }}!!!\n{% endif %}";
-  export let promptTitle = "Untitled Prompt";
-  export let paramDict: Record<string, string> = {storyTopic: "time travelling"}
-  const promptSchemaVersion: number = 2; /* 2: plain text promptText */
-                                         /* 1: HTML promptText with Jinja2 template */
+	export let promptText: string;
+  export let promptTitle: string;
+  export let paramDict: Record<string, string>;
 
   // Parse prompt args
   // const paramParseRegex = /\$\$(\w+)/gi // $$paramName
@@ -61,26 +58,6 @@
   }
   $: if (promptText, paramList, paramDict) renderedPrompt = renderPromptV1();
 
-  // Share
-  let sharedUrl: string;
-  let isSharing: boolean = false;
-  let isShared: boolean = false;
-
-  async function handleShare() {
-    isSharing = true;
-    const res = await fetch(`/api/prompt`, {
-			method: 'POST',
-			body: JSON.stringify(prompt)
-		});
-
-		const responseJson = await res.json()
-    console.log("Saved prompt with id: " + responseJson.promptId);
-    sharedUrl = $page.url.protocol + '//' + $page.url.host + '/p/' + responseJson.promptId
-    
-    isSharing = false;
-    isShared = true;
-  }
-
   import type { Editor } from "codemirror";
   
   // From https://github.com/NaokiM03/codemirror-svelte/blob/CodeMirror5/src/Codemirror.svelte
@@ -104,6 +81,7 @@
     cmTextArea.style.height = "calc( "+ cmTextArea.scrollHeight + "px - 2em)" 
     
     // TODO: fix ts(2686)
+    // @ts-ignore
     editor = CodeMirror.fromTextArea(cmTextArea, {
       mode: {name: "jinja2", htmlMode: true}
     });
@@ -142,15 +120,6 @@
     {@html renderedPrompt}
   </div>
 </div>
-
-{#if isShared}
-  <div id="sharedUrl"><a href="{sharedUrl}">{sharedUrl}</a></div>
-  <div id="sharedDisclaimer"><p>Please note that Prompter is in early development stage, your work may become altered or unreachable at any time. Use it at your own risk!</p></div>
-{:else if isSharing}
-  <button id="shareButton" class="button">saving...</button>
-{:else}
-  <button id="shareButton" class="button" on:click={handleShare}>Share</button>
-{/if}
 
 <style>
 
@@ -226,7 +195,7 @@ h2 {
   /* border: 1px solid var(--color-theme-2); */
   background: var(--color-bg-alphawhite);
   padding: 1em;
-  white-space: pre-line;
+  white-space: pre-wrap;
 }
 
 .renderError {
@@ -238,41 +207,4 @@ h2 {
   color: var(--color-theme-2);
 }
 
-/* Share */
-
-.button {
-    display:inline-block;
-    /* color:var(--color-theme-2); */
-    border:1px solid var(--color-theme-1);
-    border-radius: 2px;
-    background:var(--color-theme-1);
-    cursor:pointer;
-    vertical-align:middle;
-    max-width: 100px;
-    padding: 1em;
-    margin: 1em;
-    text-align: center;
-}
-.button:hover {
-    border:1px solid var(--color-theme-1);
-    /* background-color:white;
-    color: var(--color-theme-1); */
-    color: var(--color-bg-alphawhite);
-}
-
-#sharedUrl {
-  background: white;
-  margin: 1em;
-  padding: 1em;
-}
-
-#sharedUrl a {
-  color: var(--color-theme-2);
-}
-
-#sharedDisclaimer p {
-  font-style: italic;
-  font-size: 0.9em;
-  color: var(--color-bg-alphawhite);
-}
 </style>