+++ prompter_app/src/routes/api/chain/+server.ts
@@ -1,9 +1,8 @@
-import util from 'util';
 import { error } from '@sveltejs/kit';
 import type { RequestHandler } from '@sveltejs/kit';
 
-import type { Prompt } from '$lib/prompts';
-import { savePrompt } from '$lib/prompts';
+import type { PromptChain } from '$lib/prompts';
+import { saveChain } from '$lib/prompts';
 
 export const GET = (({ url }) => {
     return new Response(`Error`);
@@ -13,13 +12,13 @@ export const GET = (({ url }) => {
  * Create a new prompt record in database
  */
 export const POST = (async ({ request }) => {
-    const prompt: Prompt = await request.json();
-    const promptId = crypto.randomUUID(); // TODO: polyfill
+    const chain: PromptChain = await request.json();
+    const chainId = crypto.randomUUID(); // TODO: polyfill
     const editKey = crypto.randomUUID(); // TODO: polyfill
-    console.log(`POST /api/prompt. Generated promptId: ${promptId}`);
-    savePrompt(promptId, prompt, editKey);
+    console.log(`POST /api/chain. Generated chainId: ${chainId}`);
+    saveChain(chainId, chain, editKey);
     return new Response(JSON.stringify({
-        promptId: promptId,
+        chainId: chainId,
         editKey: editKey
     }));
 }) satisfies RequestHandler;