+++ prompter_app/src/routes/p/[[promptId]]/+page.server.ts
@@ -1,10 +1,10 @@
-import type { Prompt } from '$lib/prompts';
-import { loadPrompt } from '$lib/prompts';
+import type { PromptChain } from '$lib/prompts';
+import { loadChain } from '$lib/prompts';
 
 export function load({ params }) {
-	let promptRecord: Prompt = params.promptId ? loadPrompt(params.promptId) : null;
+	let promptChain: PromptChain | null = params.promptId ? loadChain(params.promptId) : null;
 
 	return {
-        promptRecord: promptRecord
+        promptChain: promptChain
 	};
 }