+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## v0.0.2 (in development)
+## v0.0.4 (in development)
+
+* Update data model to support prompt chains
+
+## v0.0.3
+
+* Add Jinja2 syntax highlighting with CodeMirror 5
+* Change prompt record version to 2 (prompt is plain text instead of HTML)
+
+## v0.0.2
 
 * Add sharing via permalink
 * Change prompt template syntax from custom to Jinja2 via [nunjucks](https://mozilla.github.io/nunjucks/)